[Update] Refactor
