+++ src/main/java/seedu/duke/Command/AttendanceCommand/DeleteStudentCommand.java
@@ -1,4 +1,7 @@
-package seedu.duke;
+package seedu.duke.Command.AttendanceCommand;
+
+import seedu.duke.*;
+import seedu.duke.Command.Command;
 
 /**
  * Class to handle deletion of Student.
