+++ src/main/java/seedu/duke/Command/AttendanceCommand/ExitCommand.java
@@ -1,4 +1,9 @@
-package seedu.duke;
+package seedu.duke.Command.AttendanceCommand;
+
+import seedu.duke.Command.Command;
+import seedu.duke.DukeException;
+import seedu.duke.Storage;
+import seedu.duke.Ui;
 
 /**
  * Class representing a command to exit the program.
