+++ src/main/java/seedu/duke/ParserStorage.java
@@ -19,9 +19,9 @@ public class ParserStorage {
         String[] studentParts = line.split("\\|");
         try {
             String studentName = studentParts[0].strip();
-            String studentAttendance = studentParts[2].strip();
             String description = studentParts[1].strip();
-            Attendance student = new Attendance(studentName,description,studentAttendance);
+            String studentAttendance = studentParts[2].strip();
+            StudentAttendance student = new StudentAttendance(studentName,description,studentAttendance);
             return student;
         } catch (Exception e) {
             throw new DukeException(CORRUPTED_TASK);
@@ -36,9 +36,10 @@ public class ParserStorage {
      */
 
     public static String toStorageString(Student student) throws DukeException {
-        if (student instanceof Attendance) {
-            return student.getStudentName() + " | " + ((Attendance) student).getDescription()
-                    + " | " + ((Attendance) student).getAttendance();
+        if (student instanceof StudentAttendance) {
+            return student.getStudentName() + " | "
+                    + ((StudentAttendance) student).getDescription() + " | "
+                    + ((StudentAttendance) student).getAttendance();
         }
         throw new DukeException(CORRUPTED_TASK);
     }
